+++ javascript/node/selenium-webdriver/http/index.js
@@ -86,7 +86,15 @@ HttpClient.prototype.send = function(httpRequest, callback) {
 var sendRequest = function(options, callback, opt_data) {
   var request = http.request(options, function(response) {
     if (response.statusCode == 302 || response.statusCode == 303) {
-      var location = url.parse(response.headers['location']);
+      try {
+        var location = url.parse(response.headers['location']);
+      } catch (ex) {
+        callback(Error(
+            'Failed to parse "Location" header for server redirect: ' +
+            ex.message + '\nResponse was: \n' +
+            new HttpResponse(response.statusCode, response.headers, '')));
+        return;
+      }
 
       if (!location.hostname) {
         location.hostname = options.host;